use concern for PII logging related methods
